+++ R/arrange.R
@@ -75,8 +75,8 @@
 #' }
 #' tidy_eval_arrange(mtcars, mpg)
 #'
-#' # use across() to access select()-style semantics
-#' iris %>% arrange(across(starts_with("Sepal")))
+#' # Use `across()` or `pick()` to select columns with tidy-select
+#' iris %>% arrange(pick(starts_with("Sepal")))
 #' iris %>% arrange(across(starts_with("Sepal"), desc))
 arrange <- function(.data, ..., .by_group = FALSE) {
   UseMethod("arrange")