+++ R/deprec-do.R
@@ -7,7 +7,7 @@
 #' felt like it belonged with the rest of dplyr. It's replaced by a combination
 #' of [summarise()] (which can now produce multiple rows and multiple columns),
 #' [nest_by()] (which creates a [rowwise] tibble of nested data),
-#' and [across()] (which allows you to access the data for the "current" group).
+#' and [pick()] (which allows you to access the data for the "current" group).
 #'
 #' @param .data a tbl
 #' @param ... Expressions to apply to each group. If named, results will be
@@ -18,11 +18,11 @@
 #' @export
 #' @examples
 #' # do() with unnamed arguments becomes summarise()
-#' # . becomes across()
+#' # . becomes pick()
 #' by_cyl <- mtcars %>% group_by(cyl)
 #' by_cyl %>% do(head(., 2))
 #' # ->
-#' by_cyl %>% summarise(head(across(), 2))
+#' by_cyl %>% summarise(head(pick(everything()), 2))
 #' by_cyl %>% slice_head(n = 2)
 #'
 #' # Can refer to variables directly