+++ R/distinct.R
@@ -47,8 +47,8 @@
 #' # You can also use distinct on computed variables
 #' distinct(df, diff = abs(x - y))
 #'
-#' # use across() to access select()-style semantics
-#' distinct(starwars, across(contains("color")))
+#' # Use `pick()` to select columns with tidy-select
+#' distinct(starwars, pick(contains("color")))
 #'
 #' # Grouping -------------------------------------------------
 #' # The same behaviour applies for grouped data frames,